refactor(go): ✨ merge files and remove Rust-isms
Several small files were dissolved into more appropriate locations.
runtime_closure.go (21 lines, only used from main) was folded into
main.go. socket_path.go and socket_framing.go were merged into
socket.go as a single cohesive socket layer. The project-file
utilities gcRootPathForProject, nixFilePathForProject, and fileExists
were moved from ops_direnv.go into communicate.go, where the
ProjectFile type lives, since they are used across many files.

Patterns carried over from the Rust transliteration that don't fit
Go idioms were also cleaned up. ReducePaths no longer uses an
intermediate reduction{} struct; survivors are built directly.
replaceWatchPath (used once) is inlined. EmptyNixOptions() is removed
in favour of the zero value NixOptions{}. The repeated os.RemoveAll
on every early return in InstantiateAndBuild and BuildFlake is
replaced by a single defer with a cleanup flag. The two identical
flakeEntry/treeEntry types are merged into one regexEntry. The
sortGCRootInfos helper is replaced with sort.Slice. parseDirenvVersion
uses fmt.Sscanf instead of a manual inner function. The three
minDirenv* constants are collapsed into one var. isMissingExec was
dead code duplicating isNotFound and is removed. The custom
stringSliceFlag type is replaced with flag.Func (available since
Go 1.16).
